Silo Scroll Menu Finish
Finished working on the machines scroll menu for the silo. Might add a feature to combine multiples of machines into one mini-popup with combined material counts.
Also thinking of having a way to toggle and see if machines are on or off from the silo menu.

Beginning to rework the player inventory to use from the silo, also hovering over span buttons will show required materials and amounts.